django_api/requirements.txt
django
opencv-python-headless
djangorestframework
markdown
django-filter
gunicorn